refactor(adapters): move adapters into separate folder
