+++ test/querybuilder.test.js
@@ -14,7 +14,7 @@
 import assert from 'assert';
 import { load as loadquerystring } from '../src/loaders/url.js';
 import { load as loadtext } from '../src/loaders/text.js';
-import { createfilter } from '../src/filter.js';
+import { adapt as createfilter } from '../src/adapters/filter.js';
 import { transformconjunctions, flat } from '../src/util.js';
 import { qb } from '../src/index.js';
